Fix crash where body contained a single matcher
For some reason if there is only one matcher in the request or response body, the Matcher was not considered as conforming to `MatchingRuleExpressible` type and it caused a non-encodable element crash. Same situation with `ExampleGeneratorExpressible` type. Same issue could occur if the body contained more than one of the same Matcher or ExampleGenerator type.