Mark testFunction as Sendable in setup
The testFunction collected from the `run` function is marked as Sendable. That imposes certain restrictions, such as disallowing the capture of non-Sendable values inside the testFunction.

The annotation was not extended to Pact’s internal functions, however. This was a touch confusing; is the annotation necessary? The answer is it must be Sendable: the `setupPactInteraction` function passes it to a `Task`, and the `Task`’s operation must be Sendable.